+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
+### Added
+
+- Admin: Allow quick copy a product via a button
+
 ### Fixed
 
 - Fix 'Customize Your Admin Menu' redirecting to the supplier specific edit page on saving.